requirements.txt
Flask==3.0.3
pillow==10.2.0
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
srt==3.5.3
ffmpeg-python==0.2.0
yt-dlp==2024.4.9
transformers==4.37.2
stable-ts==2.17.5
torchcam==0.4.0
python-dotenv
